Linux Mint best distro.
Linux Mint is a community-driven Linux distribution based on Ubuntu (in turn based on Debian), bundled with a variety of free and open-source applications. It can provide full out-of-the-box multimedia support for those who choose to include proprietary software such as multimedia codecs. Linux Mint is available with a number of desktop environments to choose from, including the default Cinnamon desktop, MATE and Xfce. Other desktop environments can be installed via APT, Synaptic, or via the custom Mint Software Manager. Linux Mint actively develops software for its operating system. Most of the development is done in Python and the source code is available on GitHub.